+++ @connect-shared/lib/attestations/agoraApi.ts
@@ -3,7 +3,7 @@ import type { Prisma } from '@charmverse/core/prisma-client';
 import { prisma } from '@charmverse/core/prisma-client';
 import { POST } from '@root/adapters/http';
 import { getAttestation } from '@root/lib/credentials/getAttestation';
-import { decodeOptimismProjectSnapshotAttestation } from '@root/lib/credentials/schemas/optimismProjectSchemas';
+import { decodeOptimismProjectSnapshotAttestation } from '@root/lib/credentials/schemas/optimismProjectUtils';
 import { optimism } from 'viem/chains';
 
 import { mapProjectToOptimism } from './mapProjectToOptimism';
