--- components/common/CreateSpaceForm/index.tsx
@@ -1,1 +1,0 @@
-export { CreateSpaceForm } from './CreateSpaceForm';
