+++ lib/credentials/external/getOnchainCredentials.ts
@@ -2,7 +2,7 @@ import type { ApolloClient } from '@apollo/client';
 import { gql } from '@apollo/client';
 import { log } from '@charmverse/core/log';
 import { prisma } from '@charmverse/core/prisma-client';
-import { arrayUtils, stringUtils } from '@charmverse/core/utilities';
+import { stringUtils } from '@charmverse/core/utilities';
 import type { SchemaDecodedItem } from '@ethereum-attestation-service/eas-sdk';
 import { isDevEnv, isStagingEnv } from '@root/config/constants';
 import { getAddress } from 'viem';
