+++ lib/credentials/getEasInstance.ts
@@ -1,4 +1,4 @@
-import { EAS, getSchemaUID } from '@ethereum-attestation-service/eas-sdk';
+import { EAS } from '@ethereum-attestation-service/eas-sdk';
 
 import type { EasSchemaChain } from './connectors';
 import { getEasConnector } from './connectors';
